feat(rvr): route public-values REVEAL through rvr log-native preflight
The base RV64 loadstore seam rejected STORED with e == PUBLIC_VALUES_AS
(the RV64 public-output REVEAL), so any program revealing public values
fell back to interpreter preflight tracegen.

- Split the RV64I loadstore registration along the enum grouping:
  zero-extension loads keep e == RV64_MEMORY_AS; stores admit
  e in {RV64_MEMORY_AS, PUBLIC_VALUES_AS} through one stored predicate
  (is_rv64_store_instruction) consulted by both routing and assembly.
- The loadstore assembler expects the store write in the instruction's
  actual address space instead of hardcoding RV64_MEMORY_AS.
- RevealInstr emits interpreter-parity preflight traces: rs1 (ptr)
  register read, then rs2 (src) register read, then a value-carrying
  block WRITE to PUBLIC_VALUES_AS via the new trace_wr_as_u64 tracer
  helper (preflight: WRITE log entry with the stored value; metered:
  record_page as before; pure/metered-cost: no-op as before). The old
  emission ticked in src-then-ptr order and logged a width-0 TOUCH,
  which could not reproduce the interpreter's records or touched
  memory.
- Add a non-vacuous reveal differential (SystemRecords equality plus a
  LoadStore AIR byte-compare selected by opcode->AIR-id, shared
  periphery excluded), registry admission assertions, and rvr-routed
  reveal end-to-end prove+verify in single- and multi-segment
  configurations.
- Fix a pre-existing u64/u32 assert_eq type mismatch that broke
  cargo test -p openvm-circuit --features rvr compilation.

+++ crates/vm/src/arch/rvr/preflight.rs
@@ -789,7 +789,7 @@ mod tests {
                 .map(|entry| entry.pc)
                 .collect::<Vec<_>>(),
             (0..program_len as u32)
-                .map(|idx| idx * DEFAULT_PC_STEP)
+                .map(|idx| u64::from(idx * DEFAULT_PC_STEP))
                 .collect::<Vec<_>>()
         );
         assert_eq!(
